python/requirements.txt
rich
typer

# gen-project-banner
Pillow
